removing api middle layer and creating server actions, refactoring some public components and admin dashboard, went from views to routes for the different tabs in admin sidebar
