+++ app/(admin)/admin/_components/EditExperiencesContent.tsx
@@ -1,12 +1,7 @@
 "use client";
 
-import {
-  deleteExperience,
-  deleteStorageFolder,
-  fetchExperiences,
-} from "@/utils/clientActions";
 import { ExperienceItem } from "@/utils/types";
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import ExistingExperienceCard from "./ExistingExperienceCard";
 import {
   Dialog,
@@ -17,37 +12,48 @@ import {
   DialogTitle,
 } from "@/components/ui/dialog";
 import ExistingExperienceForm from "./ExistingExperienceForm";
-import LoadingIndicator from "../../../../components/LoadingIndicator";
 import { Button } from "@/components/ui/button";
 import { toast } from "sonner";
+import { deleteExperience } from "@/actions/experiences";
+import { deleteStorageFolder } from "@/actions/storage";
+
+type EditExperiencesContentProps = {
+  experiences: ExperienceItem[];
+};
 
 //this will make this page dynamic and fetch for experiences on every page request
-export const dynamic = "force-dynamic";
+// export const dynamic = "force-dynamic";
 
-const AllExperiencesWrapper = () => {
-  const [allExperiences, setAllExperiences] = useState<ExperienceItem[]>([]);
+const EditExperiencesContent = ({
+  experiences,
+}: EditExperiencesContentProps) => {
   const [activeItem, setActiveItem] = useState<ExperienceItem | undefined>(
     undefined,
   );
-  const [loading, setLoading] = useState(true);
   const [isFormOpen, setIsFormOpen] = useState(false);
   const [isDeleteConfirmOpen, setIsDeleteConfirmOpen] = useState(false);
-  const [errorFetchingExp, setErrorFetchingExp] = useState("");
 
   const deleteActiveExperience = async (experience: ExperienceItem) => {
     try {
       console.log("Deleting experience: ", experience.title);
 
-      const deResult = await deleteExperience(experience.storageFolder!);
-      toast.success(<p className='text-sm'>{deResult.successMessage}</p>);
+      const isExperienceDeleted = await deleteExperience(
+        experience.storageFolder!,
+      );
+      const isStorageFolderDeleted = await deleteStorageFolder(
+        experience.storageFolder!,
+      );
 
-      const dsfResult = await deleteStorageFolder(experience.storageFolder!);
-      toast.success(<p className='text-sm'>{dsfResult.successMessage}</p>);
+      if (isExperienceDeleted && isStorageFolderDeleted) {
+        toast.success(
+          <p className='text-sm'>
+            {experience.title} was deleted successfully!
+          </p>,
+        );
+      }
 
       setIsDeleteConfirmOpen(false);
       setIsFormOpen(false);
-
-      getExperiences();
     } catch (error) {
       console.error(error);
       const errorMessage =
@@ -56,59 +62,20 @@ const AllExperiencesWrapper = () => {
     }
   };
 
-  const getExperiences = async () => {
-    setLoading(true);
-    setErrorFetchingExp("");
-
-    const result = await fetchExperiences();
-
-    if (!result.data && result.errorMessage) {
-      setErrorFetchingExp(result.errorMessage);
-    }
-
-    if (result.data) {
-      setAllExperiences(result.data);
-    }
-
-    setLoading(false);
-  };
-
-  useEffect(() => {
-    getExperiences();
-  }, []);
-
-  if (errorFetchingExp.length > 0) {
+  if (experiences.length === 0) {
     return (
-      <section className='flex flex-col justify-center items-center gap-1 w-full'>
-        <p className='font-bold text-xl'>{errorFetchingExp}</p>
-        <Button onClick={getExperiences}>Retry</Button>
-      </section>
-    );
-  }
-
-  if (loading) {
-    return (
-      <section className='mx-auto'>
-        <LoadingIndicator className='w-16 h-16' />
-      </section>
-    );
-  }
-
-  if (allExperiences.length === 0) {
-    return (
-      <section className='mx-auto'>
+      <section className='mx-auto mt-8'>
         <p className='font-bold text-xl'>No experiences found.</p>
       </section>
     );
   }
 
   return (
-    <section className='w-full'>
+    <section className='mt-8 w-full'>
       <div className='grid grid-cols-3 gap-y-6'>
-        {allExperiences.map((item) => (
+        {experiences.map((item) => (
           <ExistingExperienceCard
             data={item}
-            loading={loading}
             onClick={() => {
               setActiveItem(item);
               setIsFormOpen(true);
@@ -146,7 +113,6 @@ const AllExperiencesWrapper = () => {
           <ExistingExperienceForm
             existingExperience={activeItem!}
             closePopUp={() => setIsFormOpen(false)}
-            refreshAll={getExperiences}
           />
         </DialogContent>
       </Dialog>
@@ -175,4 +141,4 @@ const AllExperiencesWrapper = () => {
   );
 };
 
-export default AllExperiencesWrapper;
+export default EditExperiencesContent;
