+++ app/(admin)/admin/_components/ExistingExperienceCard.tsx
@@ -5,18 +5,15 @@ import { format } from "date-fns";
 import { Edit } from "lucide-react";
 import Image from "next/image";
 import React from "react";
-import LoadingIndicator from "../../../../components/LoadingIndicator";
 
 type ExistingExperienceCardProps = {
   data: ExperienceItem;
   onClick: () => void;
-  loading: boolean;
 };
 
 const ExistingExperienceCard = ({
   data,
   onClick,
-  loading,
 }: ExistingExperienceCardProps) => {
   const { title, date, flyerUrl } = data;
   const thumbnailSrc = flyerUrl === "" ? "/assets/TSWLOGO2025.png" : flyerUrl!;
@@ -25,19 +22,13 @@ const ExistingExperienceCard = ({
     <div className='flex gap-4 border border-neutral-200 rounded-md p-4 bg-neutral-50 flex-1 w-93 2xl:w-108'>
       <div className=''>
         <div className='relative h-40 w-40'>
-          {loading ? (
-            <div className='w-full h-full mx-auto flex items-center justify-center'>
-              <LoadingIndicator />
-            </div>
-          ) : (
-            <Image
-              src={thumbnailSrc}
-              alt={`${title}'s thumbnail`}
-              fill
-              sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
-              className='rounded-md border border-neutral-900 object-cover'
-            />
-          )}
+          <Image
+            src={thumbnailSrc}
+            alt={`${title}'s thumbnail`}
+            fill
+            sizes='(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw'
+            className='rounded-md border border-neutral-900 object-cover'
+          />
         </div>
       </div>
       <div className='flex flex-col justify-between'>
