+++ app/(admin)/admin/home/page.tsx
@@ -1,47 +1,27 @@
 "use client";
 
-import { useAuth } from "@/context/AuthContext";
-import { useRouter } from "next/navigation";
 import React, { useEffect, useState } from "react";
-import Sidebar from "../_components/Sidebar";
-import { adminViews, ViewKey } from "@/utils/adminSidebarLinks";
 
 const AdminHome = () => {
-  const { user, loading } = useAuth();
-  const router = useRouter();
-  const [activeView, setActiveView] = useState<ViewKey>("landing");
-  const ActiveViewComponent = adminViews[activeView].component;
+  const [quote, setQuote] = useState("");
 
   useEffect(() => {
-    if (!user && !loading) {
-      router.push("/admin");
-    }
-  }, [user, loading, router]);
+    const getQuote = async () => {
+      const response = await fetch("/api/quotes");
 
-  if (loading) {
-    return (
-      <div className='flex items-center justify-center min-h-screen'>
-        Loading...
-      </div>
-    );
-  }
+      if (response.ok) {
+        const data = await response.json();
+        setQuote(data);
+      }
+    };
 
-  if (!user) {
-    return null;
-  }
+    getQuote();
+  }, []);
 
   return (
-    <main className='flex'>
-      <Sidebar
-        activeView={activeView}
-        onViewChange={setActiveView}
-      />
-      <section
-        id='content'
-        className='h-screen flex-1 flex ml-[12rem] mr-8 lg:ml-[20rem]'
-      >
-        <ActiveViewComponent />
-      </section>
+    <main className='py-24'>
+      <h1 className='text-4xl font-bold'>Hello, Wilma 😃</h1>
+      <p className='text-[1rem] lg:text-xl italic mt-4'>{`"${quote}"`}</p>
     </main>
   );
 };
