+++ package.json
@@ -10,6 +10,7 @@
     "email": "email dev --dir ./components/emails"
   },
   "dependencies": {
+    "@google-cloud/storage": "^7.19.0",
     "@gsap/react": "^2.1.2",
     "@hookform/resolvers": "^5.2.1",
     "@radix-ui/react-checkbox": "^1.3.2",
