+++ utils/adminSidebarLinks.ts
@@ -1,20 +1,15 @@
-import AddNewExperience from "@/app/(admin)/admin/_views/AddNewExperience";
-import EditExperiences from "@/app/(admin)/admin/_views/EditExperiences";
-import Landing from "@/app/(admin)/admin/_views/Landing";
 import { Edit, Home, Plus } from "lucide-react";
 
 export const adminViews = {
-  landing: { label: "Home", component: Landing, icon: Home },
+  landing: { label: "Home", href: "/admin/home", icon: Home },
   addNewExperience: {
     label: "Add New Experience",
-    component: AddNewExperience,
+    href: "/admin/home/add-new-experience",
     icon: Plus,
   },
   editExperiences: {
     label: "Edit Experiences",
-    component: EditExperiences,
+    href: "/admin/home/edit-experiences",
     icon: Edit,
   },
 } as const;
-
-export type ViewKey = keyof typeof adminViews;
